Fairly major refactor of main objects
This commit encompasses a pretty big re-write of most of the guts of the
modules. The Filter object has been gutted and its innards moved into a
new Bio::Path::Find::Lane object, which encapsulates the DBIC lane
object and the logic for finding files for that lane. The finding logic
is untested and desperately needs a set of test files.