Set up Config::Any to interpolate env vars
Rather than embed passwords, etc. in configuration files, this commit
tweaks the set up of Config::Any so that it will set up Config::General
to do variable interpolation. That lets us use environment variables in
conf files and have them passed through to the app.

Ditched the support for YAML, which is easier than trying to work out if
it can be set up to do the same trick in Config::Any.