Remove production DB code and add Log4perl init
This commit removes the code that used to handle lists of production
databases and ordering the database names. There's no need to order the
databases for searching, so this code is redundant.

Unless the Log4perl env is initialised in the object that instantiates
this class, L4P issues a warning. To avoid the warning during testing,
add a config to this class too. It shouldn't interfere with normal
usage, since it will only be used when L4P isn't otherwise initialised.