Add framework for adding roles to generate stats
A first stab at adding functionality to gather and display statistics on
the Lane object. So far there's a "base" role that sets up slots for the
stats data and a few common methods, and a Path role that will generate
stats for the pathfind script.

The idea is that for each "*find" script that needs to generate a stats
report, there will be a separate stats role, to pulls out just the
specific data that is needed by the script. The roles are applied as
traits of the Lane when it's instantiated.