Finalise stats collection
This commit touches a lot of files but it's mostly focussed on changes
that finish off the code to collect statistics for a lane.

There's now a class, LaneStatusFile, which wraps the '*_job_status'
files that live in a lane's directory, and another class, LaneStatus,
that wrangles them. Each Lane now gets a LaneStatus object.

The code that actually collects and organises the stats for a Lane is in
a pair of Roles, Stats and Stats/Path. The Stats role has methods that
query the database and files to produce the raw numbers. The other role
holds a mapping that selects the individual data items that are required
for a given find script's needs. In the case of Stats/Path, the mapping
selects items that are required for the pathfind stats option.

The idea is that the different *find scripts will each have their own
Stats/* role with a custom mapping.