Rename Role
Renamed the role that currently handles collection of stats
(Role::Stats::Path) to be just Role::PathFind. We can use it as a role
that tunes the Find class for use with, specifically, pathfind.
Similarly, we can later add roles to adapt B::P::Find for use with
different find scripts, such as annotationfind, etc.