Delete unused test data
This commit removes a set of SQLite databases and their corresponding
directory structures. The directory trees contain an awful lot of files,
none of which are used by the current tests -- the tests only need a
single DB, pathogen_prok_track, so carrying around all of these others
is just slowing down things like git status commands for no reason.

The test files can all be found using a tag on the master branch,
"with-multiple-test-dbs".