Move tests from App::PathFind to AppRole
Most of the code in App::PathFind is boilerplate from
Bio::Path::Find::App::Role::AppRole. Move the bulk of the tests so that
we're testing the AppRole boilerplate, rather than doing it through the
app class.

Move attributes that are specific to PathFind into App::PathFind, rather
than having them in the AppRole.

Also, fix command line logging, which breaks when we move attributes out
of the role. Just build the command line using @ARGV now.