Add progress bar when adding files to tar archive
Turns out it's possible to add individual files to the tar archive, so
do that and add a progress bar.

Fix a bug that trimmed off a leading slash when showing files that have
been added to the archive. Trim a copy of the filename now, rather than
the original.