Add switch to turn off compression of tar archives
Added "--no-tar-compression | -u" to turn off gzip compression of tar
archives. This is off by default, i.e. we gzip tars by default, but it
would make sense to reverse that: the compressed tar file is essentially
the same size as the uncompressed one, since the contents are gzipped
files. Zips are still always compressed.