Remove the role and tests for environemt
Completely remove the HasEnvironment Role and its associated tests. Fix
the other classes and tests that were still relying on it.

The HasConfig role now tries to read its config file from the filename
specified by the PATHFIND_CONFIG environment variable, and defaults to
a hard-coded path otherwise. This still needs some though, because it's
not a good place to hard code anything.

Added a test to make sure that we can use a YAML config file as well as
Config::General.