Fix a stupid, stupid bug with the list separator
This commit fixes a bug that causes hard-to-debug problems downstream,
notably with using the "remove" option in Term::ProgressBar. In this
module there's a debug statement that prints the list of database names
found by interrogating the database. In order to simplify the debug
statement, the list separator is set to newline ($,="\n"), but unless
that setting is localised, it screws up all sorts of printing, including
in progress bars, elsewhere.

This commit just add a "local" to the closure where $, is set.

Stupid bug.