Don't configure logging to write to file
The logging config previously set up both screen and file appenders,
with the file appender used to write the command line as a log message
to a file on disk. We do that logging in the main script now, so the
Log4perl config doesn't need to set it up.

Remove tests related to the command line log messages from the tests.