Fix progress bars for gzipping and tar writing
The change to Term::ProgressBar::Simple broke the updating of the
progress bar for the methods that gzip and write tar archives. This
commit fixes them by setting the progress bar limit to the number of
chunks that we're using.

Also, remove the dependency on the "--zip" option. Handling of
dependencies using MooseX::App is stil weird.