Add module to build "linked" dir for tests
There are problems when using dzil to build a distribution, in that it
ignores the directory in the test suite that contains only links to the
data files (t/data/linked). Add a module that creates the links when
required by a test script. The module reads the paths that it needs from
a text file in t/data.