Make Dist::Zilla keep empty test data dirs
The default behaviour of Dist::Zilla and the @Basic bundle is to prune
out empty directories. We need a couple of empty directories for some of
the tests, so we add ".keep" files to pin them in git, then tweak the DZ
config to make sure the dot files are found.

Also fix the main module, so that an abstract is extracted correctly and
without any warnings.