Get script name-to-lane role mapping from local var
When finding lanes we apply a Role to the Lane objects that we return.
The Role can adapt the base object to suit a particular use. For now
there's only one role, which adds functionality to the Lane to allow it
to look up statistics for a given lane.

The name of the Role to apply is currently determined by looking up a
mapping in the config and using the name of the calling script ($0) as
the key. This mapping isn't really something that needs to be in the
config, since it's the code that defines what Role we need.

This commit starts to fix this by hard-coding the mapping into the class
definition and using that when determining the Role to be used.