Fix behaviour of lane_role attribute
If we're using the B::P::F::Finder class outside of the context of the
pathfind script, it doesn't make sense to throw exceptions when the
lane_role attribute is specified, since that's intended for fine-tuning
the Lane object to work with a given *find script.

Instead, this commit tweaks the behaviour of the Finder to make it
simply try to find and apply a role to the Lane objectss that it
returns, but not to get upset if it can't find a Role to apply. An
exception is still thrown if the class finds the name of a Role to apply
but the Role doesn't exist.

Tests have been updated accordingly.