Parallelise the database searching
It's a bit of a convoluted process, but this commit adds a new method,
"_find_lanes_in_parallel", which parallelises the database searches,
running the search in a separate fork for each database.

The changes here break existing tests and the new method isn't tested at
all, so this branch is highly experimental.