Handle multiple schemas
The DatabaseManager and Database classes could previously handle only
one schema, which was hard-coded as Bio::Track::Schema. When developing
infofind, we suddenly need to connect to the sequencescape warehouse
too.

That can either be done with an ad hoc connection, which doesn't seem
right, so this commit tweaks the classes that handle databases and makes
them accept a "schema_name" attribute, which tells them where to look in
the config for connection parameters. The connection params now also
include "schema_class", which specified which DBIC schema should be used
for the specified database.

These changes necessitate a few small changes in the configuration and
in the classes that use DatabaseManager. This commit fixes the affected
classes and updates tests to handle the changes.

TODO: we could really do with some new tests to make sure the new
attributes/config params are being handled correctly.