Make two attributes require Path::Class::Dir
There are a couple of attributes on B::P::F::Database that handle
directories. This commit makes them "isa PathClassDir" rathat than "isa
Str". They're mostly used internally so it's not a bit change, but
there's one use of them in tests, for the Lane object, so we have to fix
that here too.