Fail fast if possible
Add some checks to the top of the "run" method to try to fail fast when
things are otherwise likely to fail after all of the slow bits of the
process have run. For example, if we're writing statistics, that can
fail if the stats file already exists, but it will do so only after
spending time searching for lanes and collecting stats. Instead, try to
fail before doing the slow stuff.