Add perl module dependencies to tests
When setting up the distribution for travis, there was a problem with a
missing dependency during the test phase. Although the test scripts
are working with Config::General-format config files, there's no
explicit mention of Config::General as a dependency anywhere. We just
rely on Config::Any pulling in when required.

This commit (hopefully) fixes the problem by adding "use
Config::General" to the test scripts for the HasConfig role.