Change the way the -z option works with "pf data"
The "-z" option previously worked with "-a" to make the script write a
zip file. The commit changes the behaviour so that adding "-a" writes a
tar file, "-z" writes a zip file, and it's possible and allowed to add
both, making the script write both a zip file and a tar file in the same
run. I'm not sure why you'd want to do that, but hey, who am I to tell
users how to use their disk quota.