Refactor progress bar generator into a Role
The code to generate a progress bar was previously in a private method
on the B::P::F::A::PathFind class, which was inherited by the command
classes. That meant it couldn't be used in B::P::F::Finder. Refactor the
affected classes to move the method for creating a progress bar into a
Role, then apply that Role to the app class and B::P::F::Finder. Tests
are unaffected.