Get rid of the lane_role mapping
The idea of having a mapping between script name and Lane Roles breaks
down when we have a single script, "pf". This commit changes the way
that the Roles are decided on for a given command (e.g. "pf data), so
that they're now written into the command classes. So far that only
affects the B::P::F::A::P::Data command class, but the Assemby class
will also need to apply Roles to Lanes later.