Change behaviour of status file class
Previously, if, when reading a status file, we found that the specified
config file didn't exist, we'd warn and move on. That turns out to cause
problems with some lanes, so we now silently ignore the issue. If a
status file refers to a non-existent config file, we carry on
regardless.