Add code to gather stats for assemblies
Add a mechanism that allows arbitrary attributes to be set on Lane
objects as they're created by the Finder. This allows things like the
command line option for choosing an assembler to be passed down to
Lanes.

Add the start of the stats gathering code for assemblies. This is
complicated by the fact that a lane can have multiple assemblies, so the
code may need to return multiple rows of statistics for a single lane.