Add mechanism to force overwriting of existing files
Add a "--force" / "-F" option to the main PathFind app, which is a flag
to show that existing files should be overwritten. This doesn't affect
the creation of symlinks, where it's still an error to try to overwrite.

Creating symlinks is more difficult to simply force, because it the perl
built-in "symlink" refuses to create links at existing destinations, so
forcing would require us to first remove existing destination files or
directories, and I'm not sure should be doing that.