Rename Lane Roles to Lane Classes
Move things around to start converting all lane Roles into Classes. Move
test data directories accordingly.

Get rid of the data directory that used to be used by the finder tests.
The data that was in there is now in t/data/linked, so this tree isn't
needed any longer. A change to the test script (to be committed in a
little while) will remove the finally references to that directory.

Change shortcut for "no progress bars" to "N"