Add "store_filenames" attribute
Add an attribute, a boolean that affects how the Lane should store the
list of files that it finds. If "store_filenames" is false, the default,
the Lane will store found files as Path::Class::File objects, which
is what most following code expects. If "store_filenames" is true, we
store the filenames as simple strings.

The change is important for the annotation command, which hands off a
list of GFF files to Bio::AutomatedAnnotation modules. If we had to
walk the list and convert Path::Class::File objects to strings, it would
slow things down a bit, so instead we don't bother to use objects in the
first place.