Make "-p" on assembly accept multiple values
The "--program" option on the "pf assembly" command now accepts multiple
values for the assembler name. Multiple names can be added as a
comma-separated list of using multiple "-p" options.

Also, we're now getting the list of accepted assemblers from the type
library, rather than hard coding them in the class. That change is
responsible for the test changes, which are mostly just changes to the
order of lines, because the order of assemblers in the attribute has
changed -- it's now alphabetic.