Make the config a singleton
This is far too big a commit, but it's all a bit messy...

This commit makes the configuration into a singleton object, so that it
can be accessed from any class that applies the Role "HasConfig".

The code changes are fairly minor, but all over the place. The test
changes involve a bit of refactoring but mostly just the addition of a
call to clear the config every so often.

The advantage of this is that the config no longer needs to be passed
around -- the top-level script creates the application class with the
config, and then everything else can just apply the Role and call
"$self->config".

The disadvantage of this way of doing things is that the singleton
object complains if anything tries to re-initialise it, which
complicates matters in tests, where we're repeatedly creating objects
with subtly different configs. It might also get interesting if someone
wants to use bits of the codebase outside of the find scripts, but we
can burn that bridge when we come to it...