Fix problems with unreadable status files
Lane status information is stored in a file, _job_status, in a lane's
directory. Occasionally, if a pipeline step fails, for example, the
files in the directory do not have their ownership and permissions
changed so that any user can read them. In that case, a user who tries
to get statistics on that lane would previously get an exception.

This commit changes the behaviour of the status file-handling code, so
that it now issues a warning when it encounters a file that it can't
read, rather than dying.

Testing the new behaviour involves changes to the way that test data
files are handled. We can't easily commit files to git if they're not
readable, so this commit also includes a tweak to the module that
creates the "linked" directory, which copies rather than links the files
for one directory, then sets the file permissions to make it unreadable.

Also, remove the directory "t/data/linked" from git, since we generate
it on demand using the Test::Setup module.