Fix ugly terminal o/p when symlinking annotations
Previously, when creating symlinks with the annotation command, the
script would list the directory/file links as they were created, meaning
that they were interleaved with the progress bar that's shown when
creating links.

This commit tweaks the various "make_symlink" methods to collect the
list of linked directories/files and print them at the end, after the
progress bar has been removed.