Tweak the internals of Archivist for Ref command
This commit adjusts some of the internals of the Archivist Role to make
is useable with the Ref command. The main thing is to stop it assuming
that there's always a stats file to be written. That's true for the
other commands, but not for Ref, so now the Archivist::_make_* methods
check before trying to write the stats to disk and including them in the
archive.