Change default behaviour of ref command
Make the ref command return paths to fasta files for found references,
rather than paths to the directory containing all files for a reference.

Add the ability to return multiple references. There's a new option in
the interactive list of matching references, "all", which makes the
command return paths for all of the references. There's also a new
command-line option, "--all" that forces that behaviour.

Tests are woefully broken at this point...