Make RefFind warn when it can't find a given file
When the finder tries to return the path to a specific type of file, but
the expected file isn't found, issue a warning. This can happen when the
reference has a fasta file, but, say, no EMBL file. The previous
behaviour was to return the path to the directory for the reference,
along with a warning message, but that's ugly and leads to problems
downstream when handling this situation.