Don't require that extensions include a wildcard
When finding files, we use a mapping between a filetype and a regular
expression. The RE is used to select files of the specified type.
Previously, the file finding method would only try to use extensions if
the extension pattern included a wildcard.

When implementing qcfind, that restriction proves to be a problem,
because qcfind needs to retrieve files called, exactly, "kraken.report".

This commit removes the wildcard restriction.