Tweak CSV handling in main app class
The PathFind app class includes functionality for writing CSV files.
This commit adds a new private attribute to store a Text::CSV_XS object,
and makes the _write_csv method use it, rather than instantiating a
Text::CSV_XS object inside the method. This allows us to use the
attribute in qcfind and avoid a bit of duplication of code.