Make the Finder take notice of "processed"
Each lane has a "processed" value, which is a bit field showing which
pipelines have been run on that lane. The finder didn't previously use
that information, so we found a situation where a lane that had not been
important still showed up in "pf data" output.

This commit makes B::P::F::Finder aware of "processed", and adds a
filter to discard lanes where the bit field doesn't have a specified bit
set, i.e. where a lane has not yet completed the specified pipeline.