Add constants defining valid "processed" flags
Add constants to the Types module that define the valid fields in the
"processed" bit field. Also define a mapping between the friendly
pipeline name and the integer values. Finally, add a type,
"ProcessedFlag" that validates an int against the mapping, and a
coercion that can convert pipeline name to the integer value for that
pipeline.