Fix test data after adding "processed" filter
The lanes in the SQLite test database are used for tests related to
various commands, e.g. Data and Annotation. Now that we've added a
filter on the "processed" bit field, we need to tweak the database to
make sure that the lanes that we're looking for in the test scripts for,
say, the annotation command, don't get filtered out. That is, we need to
set the "annotated" bit on the "processed" bit field for those lanes
that should be found by the Annotate command.

This commit adjusts the values in the database to fix tests for the
Data, Annotation and Assembly commands. It will need further adjustment
when we come to test QC.