Fix tests after adding "processed" filters
Having added filters on the "processed" values to the various commands,
we need to fix the expected data for things like statistics, because now
we have a different set of pipelines showing up as "Done". This commit
fixes the various CSV files and embedded outputs, so that the values
match what we get with the "processed" filters in place.