Fix a bug in the linker
The linker has a mechanism for allowing lanes to edit the filenames that
are used when creating links. Unfortunately, there's a check to see if a
lane actually has the "_edit_filenames" method before trying to call it,
but that "can" call was using the wrong name.

Changed the call to use the correct name and the linking works as
expected.