Catch ugly exceptions in the main script
There are a couple of options, one for the Map command, one used by the
Annotation and Assembly commands, which should be validated by Moose and
Type::Tiny against an enum. The problem is that if a user specifies an
invalid value, rather than giving an sensible error through the
MooseX::App "you gave an invalid option" mechanism, there's an uncaught
exception that just kills the script with an ugly message.

There doesn't seem to be an easy way to fix the cause of the exception,
so the next best thing is to try to return a meaningful error message.
This commit adds a catch for the specific exception and returns
something vaguely informative instead. It's a hack but it's the best we
can do without digging into the guts of some very complex CPAN
modules...