Make stats role methods accept mapstats row
Previously the methods on the Lane::Role::Stats role would use the first
row found in the mapstats table when calculating stats. That's not
correct in most bases, since a lane will commonly have multiple mapstats
rows, with one having "is_qc" set, signifying the mapping that's
performed as part of the QC step.

This commit makes the methods that use the mapstats table accept a
mapstats row, so that a specific row can be handed in. If no row is
provided, the methods fall back on the old behaviour of using the first
row that they come to.