Fix the handling of mapstats data
There was a problem with the way the commands were using data from the
mapstats table. The old behaviour didn't correctly account for the
possibility of a mapping for QC purposes and then a second mapping to
generate an assembly. The problem was mainly visible in the stats
output.

This commit fixes the behaviour surrounding mapstats rows and adds an
attribute to lanes that decideds whether stats should be generated with
reference to the QC mapping or the non-QC mappings.

Also a few bug fixes, notably in the calculation of depth of coverage,
which now correctly follows the twisted logic of the method of the same
name in the original code.