Fix failing tests for the "qc" command
When merging pull requests into master, the test suit is run by travis,
using three different versions of perl. It turns out that the way that
perl 5.20 sorts things can be different to the way that perl 5.10 and
5.14 sorts. The different sorting behaviour manifests as test failures
on 5.20 but not 5.10/5.14.

This commit changes the contents of a kraken report in the test suite,
which contains lots of species with only a handful of reads, usually
about 1 or 2. When the summary report is generated by
Bio::Metagenomics::External::KrakenSummary, lines for species with these
tiny numbers of reads are in different orders for the different version
of perl. The fix is to remove most of the species with these low numbers
of hits, then adjust the stats that we get in the summary to match the
new reality.