Change the naming of archived files
Change the way that files in either zip or tar archives are named, so
that filenames include the ID of the lane from which the files were
taken.

Although there was already code to allow renaming of files when they're
added to an archive, the main change was making it possible for commands
classes to actually use that code, so the bulk of the changes here are
just re-wiring the Archivist Role to make that possible.