Fix a bug with the "--file_id_type" option
There was a bug in some, but not all commands, whereby they were using
the value of "--type" to give the types of files to find, even when that
type was "file". They should have been using a private attribute,
"_type", which is set by the main builder to give the actual type of the
IDs that are to be used for searching.

This commit adds an underscore to each of the affected command clases...