Get most functionality working for RNASeq command
There's a bit of gnarly wiring between the Lane parent class, the
HasMapping Role, and the Map, SNP and RNASeq Lane classes, in order to
get the file-finding working correctly with mapper/reference filtering.

This commit also tidies up a few of the other bits of functionality,
like archiving and linking. Still needs tests.